Move library sources to libplorth directory
