+++ libplorth/compiler.cpp
