+++ libplorth/context.cpp
