+++ libplorth/globals.cpp
