+++ libplorth/memory.cpp
