+++ libplorth/module.cpp
