+++ libplorth/position.cpp
