+++ libplorth/runtime.cpp
