+++ libplorth/unicode.cpp
