+++ libplorth/utils.cpp
