+++ libplorth/utils.hpp
