+++ libplorth/value-array.cpp
