+++ libplorth/value-boolean.cpp
