+++ libplorth/value-number.cpp
