+++ libplorth/value-object.cpp
