+++ libplorth/value-quote.cpp
