+++ libplorth/value-string.cpp
