+++ libplorth/value-symbol.cpp
