+++ libplorth/value-word.cpp
