+++ libplorth/value.cpp
